requirements.txt
flask-paginate
Flask-APScheduler
flask-user
flask-mail
Flask-SQLAlchemy
Flask-Session
PyMongo
TA-Lib